[sent_update] adjust `SentMessage` to bsuid changes
